removed pathed extension, moved regexp class to presentation/forms/helpers.
