--- lib/utils/extenstions/path.dart
@@ -1,3 +1,0 @@
-extension Pathed on String? {
-  String get path => '/$this';
-}
